moved Territory class to Map.cpp/.h files
