+++ Map.cpp
@@ -5,6 +5,18 @@
 #include <iostream>
 #include "Map.h"
 
+// Functions for the Territory class
+const string &Territory::getName() const {
+    return _name;
+}
+
+void Territory::setName(const string &name) {
+    _name = name;
+}
+
+Territory::Territory(const string &name) : _name(name) {}
+
+// Functions for the Map class
 Map::Map(Territory *territory) : territory(territory) {}
 
 Map::~Map() {
