+++ MapDriver.cpp
@@ -2,16 +2,16 @@
 // Created by Massimo Lopez on 2022-02-04.
 //
 
-
+#include <iostream>
 #include "Map.h"
-#include "Territory.h"
+//#include "Territory.h"
 
-int main(){
-    Territory* territory = new Territory("USA");
-    Map* map = new Map(territory);
+int main() {
+    Territory *territory = new Territory("USA");
+    Map *map = new Map(territory);
     cout << "Hello " << map->getTerritory()->getName() << endl;
 
-    delete(map);
+    delete (map);
     map = NULL;
     return 0;
 }
